Add new convergence criteria
Now have -- with varying degrees of polish --
population, 1st eigenvec (propogator), 1st eigenvec (l2 norm),
tmat frobenius, timescale.

Timescale is way off, but that might be due to lag time differences
as it was built with lag-time = 3

The signs of the first eigenvector elements aren't correct! This is
very weird. Even with hugely long trajectories