Add `reset` to convergence checker
Storing error values over time is a tricky issue. Ideally,
each convergence would just store one error value. That being said,
when making plots, having a whole history in each object is very nice.

When running multiple runs in a grid, the histories would not be
reset because the configurations are shared. The reset method
fixes that.

This has only been tested in non-parallel runs. I hope that when running
parallel, the configuration objects are copied